requirements.txt
Django>=4.0
djangorestframework
djoser
djangorestframework-simplejwt